# swagger: Speedup initialisation