I have checked how long does it takes to create an SDK instance in the Base app.

| | Before | After |
|---|---|---|
| On testnet | 4039ms | 1952ms |
| On mainnet | 3631ms | 1450ms |

So, this change makes SDK initialisation ~2 times faster. Also, it significantly reduces the amount of computations as I see from reports of Chrome profiler:

Screenshot of report taken before:
![before](https://user-images.githubusercontent.com/9007851/60121174-c9e0b400-9782-11e9-97db-d7f18619b3a3.png)
[full report](https://github.com/aeternity/aepp-sdk-js/files/3326646/before.txt)


After:
![after](https://user-images.githubusercontent.com/9007851/60121175-cbaa7780-9782-11e9-8387-fab4fbbca944.png)
[full report](https://github.com/aeternity/aepp-sdk-js/files/3326647/after.txt)

These extra computations make Base app freezing when it started or if current network changed, especially on mobile devices. You can compare https://stage-identity.aepps.com/ and https://feature-speedup-sdk.origin.aepps.com/ by yourself to see the difference.